fix issue with text overlaying on redrawing scenes
nx.js does not let us resize width and height of offscreen canvas at this time, so the initial fix was to just initialize the offscreen canvas with the WxH of 1280x720.

however, this offscreencanvas gets recycled, which causes that text overlaying issue. fix is to just clear the canvas context before updating texture